Removed unnecessary exported methods, and tweaked memBatch
